[frontend] fix selected feature for regardingOf filter
